shell.nix: add rust-src and llvm-tools-preview components
Tock requires the rust-src and llvm-tools-preview
components (extensions) to be present in the Rust sysroot used. Thus
this changes the Nix expression to also have these installed.

Because the previously used "rustChannelOfTargets" helper function
does not support passing an "extensions" parameter, this changes the
expression to use the more generic "rustChannelOf" function and
manually passes in the required parameters. It furthermore contains
some cleanup such that the rustBuild function is now a single
expression, without dependencies on other higher-scoped variables.

Signed-off-by: Leon Schuermann <leon@is.currently.online>