Restructure the USB stack to be more modular
This commit restructures the USB stack with a different division of
responsibilities and allows for more modularity. In particular, the goal
is to allow developers to define USB interfaces (funcional collections
of USB endpoints) independently of device and configuration definitions
and without knowing ahead of time the the specific endpoint, interface
and configuration "values" (USB's term for the position and name of
each). In turn, this allows constructing a USB device by combining
multiple interfaces (e.g., a device that is both a serial and HID
device, multiple serial devices, etc) as well as moving responsibility
for defining device specific characteristics (vendor and product ID, max
current draw, etc) to more appropriate location, such as the board
definition.

Previously, the functionality for attaching to the host and responding
to device enumeration was encapsulated in a library
(capsules::usb::usbc_client_ctrl) but driven an interface/function
specific driver (e.g. capsules::usb::cdc). The interface specific driver
served as the USB `Client` and called out to this library, both decided
and relied upon it's own choices for endpoint and interface values, and
was responsible for determining device and configuration specific data.
This prohibited composition (you could not successfully initialize and
use both the CDC and CTAP interfaces, nor multiple of either) and
resulted in somewhat odd artifacts, such as all devices being declared
as independently powered regardless of whether the board was itself
powered from mains or via the USB host.

This commit flips that structure. It introduces a Device struct (based
mostly on `usbc_client_ctrl`) which is the direct client of the
low-level USB driver and holds authorotative meta-data about device
specific information (max current draw, USB version numbers,
battery-powered, etc). It also introduces traits for Configuration,
Interface and structs for Endpoint which are used to form a heirarchy
for a device's functionality in a modular way, and supplied to a
particular Device instance.

A Configuration is just a collection of Interfaces with some meta-data
(as required by the USB spec), Endpoints are the actual channels of
communication and hold a buffer + meta-data information about the
endpoint (Interrupt vs Bulk, max packet size, etc), and Interfaces are
collections of Endpoints as well as the main interface for
implementation actual useful functionality.

Enumeration is handled in the Device struct, but deferred to the
supplied Configuration for enumerating interfaces, endpoint and any
ancillary descriptors (such as Class descriptors or HID reports
descriptors). It also handles allocating configuration, interface, and
endpoint numbers and routing post-enumeration requests/responses to the
approriate interface.

Finally, the commit includes two implementations of a generic serial
interface (only works on Linux AFAIK) for demonstration purposes: one
that also implements the UartData trait and is useful as a backend for
the console, and one that implements an echo service based on the
usbc_client module.