processbuffer: fix miri issue with transmuting &u8 to &Cell<u8>
Miri check for UB, and it complains that we are casting an immutable &u8
to a &Cell<u8> since the Cell will allow for mutation (even though the
code we have in process buffer will not mutable it). It is safer to
store the read only process buffer in a normal immutable &[u8] slice
than a &[Cell<u8>] since we cannot accidentally update the data in
which we have an immutable reference.

Change-Id: I62e276f7365ee9100b56d1e3b42e13574a927f2b