hil/time: remove 'a lifetime constraint on timer
Remove the overly restrictive 'a lifetime constraint on the self
parameter for methods on the timer hil trait. This was due to needing to
insert the virtual timer into the linked list when starting timer.

Instead we follow the pattern for spi and virtual alarm: we must call
the new setup() function after creating a virtual timer.

Change-Id: Iec2c1804487d8dd1a7fcabed10d91c115f7bd1b9