grant: repurpose UpcallTable as KernelManagedData
Add access to process buffers on the second parameter of the app enter
closure. This parameter will give access to upcalls (existing behavior)
and also read-only/read-write process buffers with a lifetime limited to
the app enter closures.

This is one of many steps on the way for the allow_readonly and
allow_readwrite functions on SyscallDriver to be removed and replaced
with common kernel logic -- similar to what was done with subscribe.

Change-Id: Ia076e35929b62d9389fd57ae9d27a9d42215d518