Fix command 0 documentation: it is an existence check, and always returns Success.
LowLevelDebug called it a "driver check", which can be interpreted as indicating that it should verify the driver is working. However, command 0 simply indicates LowLevelDebug is present, not necessarily that it is working. This is consistent with TRD 104.

Also, a couple other capsules are documented as returning `0` in response to command 0, which implies they return Success with u32, but they all return Success with no data.