kernel/ipc: fix null-callback documentation in notify commands
The `Process::enqueue_task` function does not provide any information
on whether or not a scheduled callback is a null-callback.
Furthermore, we are scheduling an IPC task, which will ultimately get
routed back to us. This will always succeed as long as there is
sufficient space in the Process task queue and there are no other
kernel-internal errors. Hence this removes a misleading comments about
being able to handle null-callbacks differently here.

Signed-off-by: Leon Schuermann <leon@is.currently.online>