chips/earlgrey: Don't use MIP::MTIMER bit
The RISC-V specification provides the MIP:MTIMER bit to indicate when a
timer interrupt occurs. OpenTitan also provides a RVTIMERTIMEREXPIRED0_0
interrupt for the same functionality.

We should be only using one of the timer interrupt sources, currently we
report the MIP::MTIMER bit, but never act on it, instead handling the
interrupt generated from RVTIMERTIMEREXPIRED0_0.

This patch removes all of the MIP::MTIMER functionality and disables the
interrupt in MIE to ensure that we don't get double handle a timer
interrupt. We now only use the RVTIMERTIMEREXPIRED0_0 interrupt as
passed from the PLIC.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>