libraries/tickv: Return at least part of the value
If the total length is too long for the buffer, let's at least return
some of the data and an error. This is really useful when trying to
check is a caller has permission to delete a key, as now we don't need a
really large buffer to store the value.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>