stm32f4xx: Dma - Shared Stream Dma1&Dma2 struct
To reduce the code duplication between Dma1&Dma2, the Stream class is
moved into the shared Dma module.
Dma1 and Dma2 then implement the new StreamServer trait, that allows
them to serve their own Streams.

A further refactoring is achieved by adding a StreamPeripheral trait,
which is implmeneted in Dma1Peripheral & Dma2Peripheral and returns all
the constants that might be different between DMA peripherals.
This way the dma::Stream never has to match over the peripheral, but
only ever over to stream to access the correct registers. The data to
write into the registers will be provided by the Peripheral.
This should also make adding new peripherals cleaner, as the Stream
doesn't need to be touched, but only the Peripheral enum.