opentitan: Makefile: fixup test flash-layout
This fixes a bug, where if an application is run (i.e qemu-app) after
`make test`, the incorrect flash layout is loaded (in layout.ld). Result
of `make test` copying the test layout over and this not being cleared
when an app is loaded, thus resulting in a hang after OT is initialized.

Is fixed by removing `layout.ld` after a test is run with no additional cost to
build time.

edit: Applied fix for nexysvideo Makefile

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>