make: move -L to build.rs
Per the comment for `RUSTC_FLAGS_FOR_BIN`:

> Ultimately, this should move to the Cargo.toml, for example when
> https://github.com/rust-lang/cargo/pull/7811 is merged into Cargo.

this moves setting the linker script search path to build.rs, which is
enabled by the addition of the `cargo:rustc-link-arg=FLAG` instruction
for build scripts.

This removes the need to special case the -L flag so that it is only
used when creating a binary and not when compiling dependencies, since
that is exactly what the new build script instruction handles.