kernel: hil: symmetric_encryption: Add support for AES GCM
AES GCM is a relativley new mode of AES, formalised by NIST in 2007. It
is becoming widley adopted due to it's performance improvements.

GCM is now used in a variety of IEEE standards, including 702.11ad, as
well as SSH, TLS and OpenVPN.

Due to it's growing popularity let's add support for it to Tock.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>