kernel/processbuffer: migrate ProcessSlices towards raw pointers
This commit migrates the ProcessSlice interface in Tock to work on raw
pointers instead of slices of `Cell`s. Most notably, this prevents
usage of Rust's `Index` operator, which has been replaced by a
`ProcessSliceIndex` trait and `get` method. Furthermore, this commit
addresses

- removal of all panicing methods in relation to process buffers /
  process slices,

- documentation of the new safety requirements and invariants for the
  raw-pointer based process slice interface,

- revision of previous documentation, alignment to 80 char line width,

- adaption of capsules to this new interface, in the form of a
  straightforward translation. This adds a substantial amount of
  `unwrap()` calls in capsules, and does not apply trivial cleanups
  such as converting manually implemented for-loops for copying
  between slices with the appropriate `copy_from_slice` /
  `copy_to_slice` calls. This is done for easy validation of the
  proposed changes. Subsequent commits should clean up this code, and
  demonstrate an implementation with reduced panics on select
  capsules.

All of the above are deeply interwoven, thus they could not be
separated into different commits.

Motivation
----------

In short, the previous interface was unsound. Because fundamentally
processes may allow overlapping memory regions to the kernel,
and (mutable) aliasing of memory is unsound when used with Rust
primitives (such as `&[u8]` slice references), the previous approach
used a construct which allows mutable aliasing of memory: instead of
holding onto a slice of bytes, it provided a slice of `Cell`s of
bytes. Containing an `UnsafeCell`, these constructs allow mutable
aliasing (that is, holding multiple mutable references to their
contents simultaneously).

This approach has two issues: because `Cell` is a construct allowing
mutation of the underlying value, it is unsound to use it over
hardware-immutable memory. The exact consequences of writing to
read-only memory are implementation dependant. For potentially
hardware read-only memory, the previous implementation used a wrapper
around `Cell`, exposing only a subset of the operations possible
through `Cell`, explicitly excluding operations which mutate the
underlying value. Nonetheless, wrapping such memory in `Cell` is
unsound.

Furthermore, it is desirable to be able to treat immutable Rust
slices (`&[u8]`) as `ReadableProcessBuffer`s to pass both through
common interfaces. However, transmuting an immutable `&[u8]` Rust
slice into `&[Cell<u8>]` is an unsound operation, for the same reasons
as outlined above. For more information, refer to [1].

Approach
--------

This commit changes the underlying structure of `ReadableProcessSlice`
and `WriteableProcessSlice` to no longer be a `#[repr(transparent)]`
wrapper over a slice of `Cell`s, but instead store a non-null pointer
and a length to describe a given shared memory region. This is
fundamentally identical to the in-memory representation of slice
references. However, based on this information, raw pointer-based
methods can be used to dereference the covered memory, for which
Rust's aliasing rules do not apply as long as they can not access
other Rust allocations.

At the same time, it introduces a significant amount of unsafe code,
as many convenience methods of Rust slices can no longer be used. Thus
this code requires especially careful vetting.

[1]: https://github.com/tock/tock/issues/2882