chips: move arch config to function level
In tock we use cfg[]s based on architecture to hide assembly code when
cargo builds tests. For consistency and clarity this PR moves a couple
cfgs to the function level so we are not modifying the body of functions
using compiler flags. Also, if there is some use case for running
chip-specific code on host this will make it clear when certain
functions are unimplemented.