Console: Print the first N bytes of the write buffer, plus doc fixes.
The console syscall driver's documentation did not say what happens when the number of bytes to write (which is passed to Command) is smaller than the size of the passed buffer. I think everyone reading the documentation assumed that console would write the first N bytes of the buffer, but it actually wrote the last N bytes. This PR alters console's behavior to print the first N bytes, and states that in the documentation.

While editing the documentation, I made a few other fixes:

1. It no longer has a paragraph about buffers being automatically released after use -- that behavior was specific to Tock 1.0.
2. I clarified that console will never perform partial writes. This is to clear up the ambiguity expressed at https://github.com/tock/libtock-rs/pull/395#discussion_r824693201.