mpu: Load fixed-size apps in the middle of memory regions if needed
Fixed-size processes have a hard-coded RAM placement. Before this hange, they would be rejected if the start placement did not lie on a MPU access boundary.

In the case of cortex-m, the boundary is dependent on total RAM size, and that, in turn, on the stack size declared in the tbf header. As a result, changing stack size could turn a working application into a permanently rejected one.

This change adds some padding before the actual application start, to align the memory region with the application size.