Console: terminate write rather than panic if the Allow buffer is unexpectedly shortened.
Fixes #3007

With this PR, if a process starts a console write then replaces the write buffer with a shorter buffer, the Console capsule will terminate the write rather than panicing.

I removed the Console documentation's description of Console's likely behavior when handling a shortened buffer. It is no longer accurate, and is not relevant to users of Console because the actual behavior is explicitly undefined.

This PR was tested by using a `libtock-rs` app that starts a large console write then replaces the buffer with an empty hardware. I was able to reproduce the bug on an NRF52840-DK, and have confirmed this change fixes the panic.