Console: Add a way to cancel ongoing console writes.
`libtock-rs`'s panic handlers want to print a panic message regardless of whether or not the panic occurred with a write in progress. Printing the panic message without corruption requires a way to wait for the console to become available, and the console API does not currently have a documented way to do so.

This PR adds a way to cancel any ongoing transactions to the stable console API. It does so by documenting an exiting behavior: that calling the "start write" command with a length of `0` bytes will cancel ongoing transactions and trigger an upcall.

An alternative possibility is to add a new "cancel write" command to the API.