Wrap VolatileCell around Cell; save the world
Uses Cell as a wrapper around the value stored in a VolatileCell (and
thus inheriting an UnsafeCell). This prevents triggering undefined
behavior when sharing a VolatileCell immutably when, internally, it
might change.

A consequence is that we have to remove the Copy marker, but that was
only used in one place to make initializing a large array easier, so I
changed that call site to simply initialize the array explicitly. More
source code, no more compiled code.