doc/reference/trd-uart: disable bulk transfers with width >= 9
The TRD mentions that `transmit_buffer` and `receive_buffer` are not
supported for transfers with data-width configured to larger than
8-bit, but does not explicitly document what happens when these
methods are invoked in this configuration. This changes the text to
enforce these methods must return `ErrorCode::INVAL` in such a
configuration.

It furthermore extends section 3 (`Configuration` and `Configure`) to
describe both the case of Width < 8-bit and Width > 8-bit.

Signed-off-by: Leon Schuermann <leon@is.currently.online>