doc/reference/trd-uart: apply config changes only to new operations
It does not seem like a good idea to allow UART configuration changes
mid-flight, and conceivably some hardware might not support this at
all (e.g. fix parameters on the start of the DMA operation or raise an
exception when changing them mid-flight).

To unify the behavior of different implementations, it seems most
sensible to define that configuration changes apply only to new
operations and a device is free to refuse changing its configuration
when an operation is currently ongoing.

Signed-off-by: Leon Schuermann <leon@is.currently.online>