doc/reference/trd-uart: ensure configuration changes are atomic
The utility of `Configure::configure` is quite limited currently,
considering that other methods in the `Configure` allow setting all
parameters contained in the passed `Parameters` struct already. What
would make it significantly more useful is to have all changes applied
through that interface be atomic: if only a single requested parameter
is not supported by the underlying device, do not apply any of the
passed parameters. One particularly convincing usecase for this might
be a parameter change following a device-negotation: either the
negotiated parameters can be used, or communication can continue in
the original configuration.

Signed-off-by: Leon Schuermann <leon@is.currently.online>