doc/reference/trd-uart: return an error on {tx,rx}_len == 0
If transmissions or receptions of zero length must be supported as
proper operations, they will need to result in a callback which must
be generated as part of a different call stack. If a UART device does
not support generating hardware interrupts for the completion of
zero-length transmissions, a device driver implementation has to use a
deferred call mechanism in order to be compliant with the API
specified in this HIL.

To avoid employing a deferred call for this edge case when it would
otherwise not be required, avoid supporting zero-length operations.

Signed-off-by: Leon Schuermann <leon@is.currently.online>