Update software SHA256 implementation to new Digest HIL.
Add accessors to LeasableBufferDynamic to simplify software implementations,
such that the match can be inside the buffer type. This is a size/
speed tradeoff. Putting the match inside the accessor is slower
(match on each access) but it means there isn't duplicated code above.
Putting the match in the caller means there's only one match,
but the code inside the match might be duplicated across immutable
and mutable leasable buffers.