esp32c3: fixup Makefile bug
With the latest version of esptool (devel branch, 4.2-dev as of
19/06/22), using `elf2image` wihtout `--dont-append-digest` causes the
`--flash_mode dio` command to be ignored when flashing. As a result,
despite the flash completing, it defaults to flashing in qio, which
causes a bootloop. It seems that for now, the binary must be flashed
with dio.

This patch adds the `--dont-append-digest` where appropriate so that
flash mode is not set to default (`qio`) and the correct `dio` mode is
used.

Signed-off-by: Wilfred Mallawa <thulith.mallawa@uqconnect.edu.au>