Makefile.common: Expose virtual-function-elimination
By enabling virtual-function-elimination we can reduce the size of the
final binaries. Here is a before and after for OpenTitan:

Before:
    Binary: 139K
    .text: 22818
After
    Binary: 136K
    .text: 21d0c

As virtual-function-elimination is still experimental let's not enable
it by default, but leave it as an option. This allows people to test it
and help contribute improvements and bug reports back to the Rust
compiler.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>