build: change `CI` to `NOWARNINGS`
Tock uses different flags in CI so that warnings throw errors, while
during normal development warnings are OK. We used the CI env variable to
control this, however, this conflates Tock's policy on warnings in CI
with running in _any_ CI. That is, a fork of Tock also ends up using
Tock's CI policy on warnings, because CI tools tend to set the $CI
variable as well.

This patch renames our variable to be more clear of what our use is. In
theory we can dissallow warnings while not being in CI.