Remove `#![feature(const_mut_refs)]`
This commit removes the const_mut_refs unstable feature and
all uses of it. For the most part, this just required making every const
constructor that created a `TakeCell` no longer const, as a result of
`TakeCell::empty()` no longer being const. Thanks to the updated
peripheral instantation approach in
https://github.com/tock/tock/pull/2069 and related PRs, this was very
straightforward to perform, as peripherals no longer need to be created
in const context.