tock_registers/test_fields: respect struct size padding w/ alignment
As illustrated in tock/tock#3019, Rust causes a struct's size to be
padded to a multiple of the largest field's alignment
constraints.

The tock-register-interface crate contains some sanity checks which
ensure that a struct is not missing any padding fields, all fields are
generally well aligned, and that the fields are placed at their
respective specified memory address. As of tock/tock#2922, these
checks are no longer runtime tests, but static const assertions which
cause proper compiler errors, should an assertion be violated.

Because the tock-registers-interface macros check invariants by
iterating over the individual fields and accumulating their size, the
above property of Rust structs has not been taken into account
properly and caused compiler errors for register structs which contain
fields of varying alignment and size.

This commit changes the iterative processing of register fields to
also track the highest alignment constraint of any field encountered
and, at the end of the struct, add the remaining alignment to the
expected size of the struct. Thus the following register definition
works now:

```rust
register_structs! {
    pub RegisterBlock {
	(0x00 => foo: ReadWrite<u64>),
	(0x08 => bar: ReadWrite<u32>),
	(0x0C => @END),
    }
}

fn test_struct_loc() {
    // Undefined behavior here, but sufficient for checking offsets
    let regs = unsafe { &*(0x0 as *const RegisterBlock) };
    assert!(regs as *const _ as usize == 0x0);
    assert!(&regs.foo as *const _ as usize == 0x0);
    assert!(&regs.bar as *const _ as usize == 0x8);
}
```

Fixes tock/tock#3019.

Signed-off-by: Leon Schuermann <leon@is.currently.online>