libraries/tickv: crc32: Cleanup the CRC32 interface
The original interface was clunky and required creating a Crc struct and
then getting a Digest from there. This patch instead entirley hides the
Crc struct and we just provide a Crc32 struct for performing CRC
operations.

This patch also documents all of the public functions and structs.

Finally we also reset the value after calling finalise() to allow
reusing the Crc after it has completed.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>