time: fix secondary alarm from firing immediately during alarm callback
If you have two alarms and one sets the other in its callback, and the
second alarm is later in the linked list of virtual alarms, then the
second alarm callback will fire immediately.

We need get the latest "now" when comparing the reference time while
going through the list of virtual alarms, otherwise the assumption
that reference is always in the past breaks down and falsely
triggers alarms set during previous alarms callback.

Change-Id: I5b370f2ece06deaede2c3f4ffe42d0dae51f7f6c