boards/redboard_artemis_nano: Fixup app loading
Ensure the app doesn't override Tock.

The ROM seems to store some temporary data at the load-address-blob
address. In commit f8ceb22352da we increased the addresses, but didn't
correctly increase the load-address-blob of the application enough. This
means that the temporary data can override the Tock binary. Increase the
address to a larger one that works.

This is all a bit magic as there is no documentation on what these
addresses actually are or how the ROM loads the binary, but this works
so let's go with this.

Fixes: f8ceb22352da "boards/redboard_artemis_nano: Increase the blob loading address"
Signed-off-by: Alistair Francis <alistair.francis@wdc.com>