arm/cortex-m/syscall: no volatile userspace RAM accesses
Given the changes introduced in the previous commits (starting
51274855a "arch/cortex-m: introduce CortexMVariant trait for
arch-specific fns"), which cause the Rust compiler to have direct
visibility of the `#[naked]` functions and `asm!` macros used in
handling interrupts and switching to userspace, it is obliged to
insert a compiler fence before executing these blocks which
potentially modify userspace memory.

Hence, volatile accesses to these memory regions are not actually
needed, given that we only have a single thread of execution and
retain total ordering of accesses in hardware.

This commit thus changes volatile memory accesses for regular accesses
through pointers.

Signed-off-by: Leon Schuermann <leon@is.currently.online>