boards/opentitan: add flash_ctl utests
Adds a new test suite to test the flash controller directly.
Previous testing was implicit through the TickV tests.

This is usefull for testing OT changes to the flash controller
directly, and much faster for when running on Verilator compared to
TicKV tests.

We currently test all supported functionality,
`read_page`, `write_page` and `erase_page`.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>