redboard: add notice of 2.1 workaround
This was discussed on the Core Team call Aug 26, 2022. This is believed
to be a symptom of incorrect context switching logic in the core kernel
(which is causing userspace stack corruption). This workaround prevents
crashes on boot by eliminating function calls (and more importantly,
returns using the corrupted stack). The corruption is a transient problem
which is corrected during early boot, allowing the board to continue
normally after early boot.

Tock 2.2 intends to include a rewrite of cortex-M context switching which
should address the underlying issue.