chips/earlgrey/aes: Fixup AES for Verilator
On the latest commit of OpenTitan supported by Tock, AES does not work
as per the tests. The driver returns `ErrorCode::BUSY` since the hw_ip
has not moved into the `idle` state when we expect. This issue only occurs on Verilator
and works as expected on HW.

This patch adds a Verilator specific change to make sure we add a small
wait where appropriate. Now passing AES tests on Verilator.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>