ESP32-c3 build fixes, add docs for flashing apps
The following warning from esptool when running `make flash-app` causes all apps flashed to the esp32 board to fail:
```
Warning: Image file at 0x0 is protected with a hash checksum, so not changing the flash mode setting. Use the --flash_mode=keep option instead of --flash_mode=dio in order to remove this warning, or use the --dont-append-digest option for the elf2image command in order to generate an image file without a hash checksum
```

The flashed apps seem to overwrite critical regions of flash, causing the board to enter a infinite loop of trying to boot tock and failing.

Solution: add --dont-append-digest as the warning suggests, resolves issues