kernel: convert static_buf!() to be like singleton
That is, `static_buf!()` will now panic if the same static buffer is
used multiple times. This happens if the same `static_buf!()` expansion
is run multiple times, such is if `static_buf!()` is in a function which
is called multiple times. Since doing this causes memory to be aliased
which is bad, we panic to let the board author know this needs to be
fixed.

Also: doc: update soundness with new static_init impl