virtual_alarm: fix issue with missed alarms
Fix issue that occurs if a virtual alarm expires *after* the virtualizer
loops over it but *before* the virtualizer calculates the next alarm.
It is possible that there is an alarm that has not fired but is in the
past. The bug is that when calculating the next alarm, the virtualizer
performs a simple `wrapping_sub` of reference+dt to compare it with
`now`. This calculation will make alarm in the past appear to be in the
far future.

Therefore, when calculating what the "next" alarm is, the code needs to
consider if an alarm is in the past (but has not fired). Such an alarm
should be one of the "newest" alarms, and so cause the virtualizer to
schedule a near-immediate alarm to fire and trigger it.

Also add unit test for verify this is fixed and prevent future
regressions.

Change-Id: Ib9e95e0c2d8791b0eb544881b604ede79b6fdb51