This addresses the bug noted in #3277. The bug relates to what happens if a virtual alarm expires *after* the virtualizer loops over it but *before* the virtualizer calculates the next alarm. It is possible that there is an alarm that has not fired but is in the past. The bug is that when calculating the next alarm, the virtualizer performs a simple `wrapping_sub` of reference+dt to compare it with `now`. This calculation will make alarm in the past appear to be in the far future.
Therefore, when calculating what the "next" alarm is, the code needs to
consider if an alarm is in the past (but has not fired). Such an alarm
should be one of the "newest" alarms, and so cause the virtualizer to
schedule a near-immediate alarm to fire and trigger it.

This PR also fixes a syntax error in the macro for a multi-alarm test.