Fix MapCell safety
MapCell is unsafe as it does not distinguish between an empty cell and a
full cell that is borrowed.
It also fails to drop its contents if put is called.
I have fixed these issues and added in a more convenient try_borrow_mut
to bring the interface closer in line with RefCell.

Change-Id: I7fb258e3cfafbc451f9683e893262ba894ac8694