chips/sifive/clint: parametrize over Clint timebase frequency
It turns out that at least the QEMU emulated boards (hifive1 and
qemu_rv32_virt) provide a Clint timebase of 10MHz, whereas the
previous implementation assumed a timebase of 32KHz. This forwards the
frequency as a generic parameter over the Clint, such that the driver
implementation can be shared between chips but the frequency adjusted
to meet the actual underlying hardware. It furthermore corrects the
qemu_rv32_virt board to use the proper 10MHz frequency.

Signed-off-by: Leon Schuermann <leon@is.currently.online>