Add the required `.` at the end of copyright lines.
When I implemented the license checker in #3345, I did not notice that `trd-legal.md` required a period (`.`) at the end of the copyright line. This PR adds that period to the license headers from #3345.

This begs the question: should I add logic to the license checker than checks the format of `Copyright` lines? I'm leaning towards "no" because they *should* be mostly copy-and-paste and it minor punctuation errors don't seem very significant.