`licensecheck`: Allow an empty comment line after the header.
`trd-legal.md` says:

> License and copyright information SHOULD have at least one (1) blank line separating it from any other content in the file.

When I implemented the license checker, I interpreted that as disallowing an empty comment. However, a few files in this repository have an empty comment line after the header instead. This changes the license checker to accept either a truly blank line or an empty comment line after the license header.