Updated `insert_byte` method
Instead of checking if the buffer is filled up, try to get the byte from the desired position. If returns Some(byte_buffer) update the byte from `EOL` to `byte`, if None returned do nothing. (P.S: the same functionality is achieved).

Tested on `nrf52840dk` and works as expected.