arch/rv32i: clobber kernel registers instead of manual (un)stacking
It turns out that Rust allows us to mark the majority of kernel
registers as clobbered in an `asm!()` block on `rv32i`. We can use
this in the context switch code to drastically reduce the number of
registers we need to manually stack and unstack, and instead let Rust
do this work for us.

In the worst case, this simply means that we have less lines of asm,
and thus less things for us to break / a slightly more readable
context switch handler. It is not any more or less efficient for Rust
to stack and unstack registers before and after this `asm!()` block,
then it is for us to do it ourselves.

However, it turns out that this change allows Rust to actually only
stack the registers which need to be restored in practice, according
to the surrounding function's ABI. Thinking in terms of
caller/callee-saved registers is a bit tricky here, given that Rust
doesn't make any promises concerning the ABI of these functions (not
tagged as e.g. `extern "C"`), in practice we can expect quite a few
instructions to be saved here.

From some initial, very unscientific tests (aka. looking at the
generated assembly) it seems that we could save a good ~55
instructions per context switch in the best case here. Also, this does
seem to come with some code size savings (about 100 bytes on the LiteX
sim board).

Further investigation is necessary as to the true runtime impact &
correctness of this, but it seems to work in initial tests.

Signed-off-by: Leon Schuermann <leon@is.currently.online>