Add license headers to every file the license checker supports.
This PR adds license headers for every file type in the repository that the license checker supports. This uses the same license notice style as #3317, setting the copyright year to the current year (2023) across-the-board.

I removed the `/.envrc` and `/.direnv` exclusions, as I don't think they belong it `.lcignore`. I think they belong in `.gitignore` instead? I'd appreciate input on where they belong (or why I was wrong and they belong in `.lcignore`).

I also made a stylistic decision to have a single `.lcignore` file in the root of the repository. To that end, I kept `doc/`-specific rules in the root `.lcignore`, and removed `tools/license-checker/testdata/.lcignore`.