chips: apollo3: Set a second timer for alarms
From the Apollo3 errata list there is an entry:

"ERR026: STIMER: Glitch on XTAL clock input may cause missed compare"

that describes some alarms not triggering interrupts due to jitter in
the clock.

The workaround is to set a second alarm 1 tick later, to ensure that at
least one of them will trigger.

This patch adds that second alarm as I was seeing timer interrupts be
missed on some boards. This fixes the issue

Signed-off-by: Alistair Francis <alistair@alistair23.me>