Use a fallback syntect syntax rather than a custom Rust parser in the license checker.
The license checker uses the `syntect` crate to parse files wherever possible. `syntect` provides a number of syntax definitions for common language types. However, we have some files that `syntect` does not recognize. Currently, the license checker falls back to a hand-coded parser if it encounters a file type that `syntect` does not recognize. This PR removes that hand-coded parser, and replaces it with a fallback `syntect` syntax. This removes the duplicate parsing logic, replacing it with an easier-to-extend Sublime syntax definition file.

This will make it easier to resolve #3417, as one of the file types that `syntect` does not recognize does not support line comments (linker scripts), and parsing block comments is harder than parsing line comments.