Add support for block comments to the license checker.
As noted in #3417, there are two file types in the Tock repository that do not support line comments: XML, and linker scripts. This PR allows the license checker to operate on those files.

**Limitation**

This PR does not support the following comment format:

```rust
/* Licensed under ...
 * SPDX-License-Identifier: ...
 * Copyright ...
 */
```

because the license checker decodes the comment body as:

```
Licensed under ...
* SPDX-License-Identifier: ...
* Copyright ...
```

and errors on the leading `*`'s. I don't see an easy way to identify that comment style and filter out the leading `*`'s, and because block comments are rare in Tock's codebase, I don't feel that we should add complexity to handle them. Instead, use the following format:

```rust
/* Licensed under ...           */
/* SPDX-License-Identifier: ... */
/* Copyright ...                */
```

When I write documentation for the license checker (tracked in #3419), I will document this clearly.