capsules: spi_controller: Don't provide read buffer for write ops
The SPI HIL uses a single `read_write_bytes()` function which is used for
both reading/writing. The `read_buffer` is optional if the caller just
wants to write data and doesn't care about reading any data back they
can provide `None`

The current userspace capsule always supplies a read buffer, even if
userspace doesn't want to perform a read. This results in the data being
read from the bus, passed back to the capsule and then not being used.

To avoid reading and copying data that was never requested, let's avoid
passing the read buffer to the HIL implemenation if userspace doesn't
want to read the data. This way implementations can avoid copying the
data from the bus when it isn't going to be used.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>