tickv: async: flip err/ok
Currently the async tickv interface returns ok when things are very
wrong and err when things are working as expected. This is reasonable
inside of the library, but pushes the complexity of using a synchronous
library asynchrousnly outside of the async wrapper.

This commit flips the errors so that when a tickv function schedules an
async operation as expect it returns Ok() and only on a true error
returns Err().