boards: microbit: use helper function in main
By outsourcing all of `main()` to a separate function (in this case
`start()` other than running the kernel, we can minimize the stack frame
size of main. This is important because the main stack frame is never
popped since main does not return.

This is an alternative to doing this piecemeal for individual functions.